[UPDATE] Split up data store from settings
