+++ engine/renderer/public/renderer.hpp
@@ -3,6 +3,7 @@
 #include "application_module.hpp"
 #include "bloom_settings.hpp"
 #include "cpu_resources.hpp"
+#include "data_store.hpp"
 #include "ecs_module.hpp"
 #include "settings.hpp"
 #include "swap_chain.hpp"
